Move all tests files to __test__ folder
