+++ __test__/dropdown.spec.ts
@@ -1,7 +1,7 @@
 import 'jest';
 import '@testing-library/jest-dom/extend-expect';
 import { render, cleanup, fireEvent } from '@testing-library/svelte';
-import Dropdown from './dropdown.svelte';
+import Dropdown from '../dropdown.svelte';
 
 describe('Dropdown Test Suite', () => {
   it('Dropdown may be colorized', () => {
