+++ __test__/pagination.spec.ts
@@ -1,7 +1,7 @@
 import 'jest';
 import '@testing-library/jest-dom/extend-expect';
 import { render, cleanup, fireEvent } from '@testing-library/svelte';
-import Pagination from './pagination.svelte';
+import Pagination from '../pagination.svelte';
 
 describe('Pagination Test Suite', () => {
   it('Pagination may be colorized', () => {
